requirements.txt
# ================================================
# Consultant Copilot - Dependencies
# ================================================

# Core
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# Telegram
python-telegram-bot==20.7
telethon==1.34.0

# LLM
anthropic==0.18.0
openai==1.12.0

# RAG (Этап 2)
langchain==0.1.5
langchain-community==0.0.17
langchain-openai==0.0.5

# Todoist
todoist-api-python==2.1.3

# Notion (для миграции)
notion-client==2.2.1

# Utils
aiofiles==23.2.1
python-multipart==0.0.6
tenacity==8.2.3

# Dev
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
